% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_zarr.R
\name{install_zarr}
\alias{install_zarr}
\title{Helper function to install the `zarr` python module}
\usage{
install_zarr(envname = "ZarrExperiment")
}
\arguments{
\item{envname}{`character(1)` virtual environment in which to
install the python zarr module.}
}
\value{
Reference to the python module, invisibly.
}
\description{
Helper function to install the `zarr` python module
}
