% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZarrRemote-class.R
\name{ZarrRemote}
\alias{ZarrRemote}
\title{A remote S3-like representation of Zarr folders}
\usage{
ZarrRemote(endpoint, bucket, resource, ...)
}
\arguments{
\item{endpoint}{character(1) The base URL of the remote Zarr folders}

\item{bucket}{character(1) The bucket name portion and any subdirectories in
the full URL.}

\item{resource}{character(1) Optional. A full URL typically composed of both
`endpoint` and `bucket` components.}
}
\description{
The `ZarrRemote` class allows one to display and import Zarr
  files on the web that are accessible via the `S3Fs` python module
  <https://s3fs.readthedocs.io/en/latest/> and based on the AWS S3 service.
}
\details{
The `endpoint` component consists of both the `scheme` and the
  `domain`. The complete URL is the combination of the `endpoint` and
  `bucket` components. A complete URL can also be provided via the `resource`
  argument.
}
\examples{

zr <- ZarrRemote(
    endpoint = "https://mghp.osn.xsede.org/",
    bucket = "bir190004-bucket01/TMA11/zarr/"
)

zr <- ZarrRemote(
    resource = "https://mghp.osn.xsede.org/bir190004-bucket01/TMA11/zarr/"
)

zr

\dontrun{
  zarray <- import(zr, filename = "5.zarr")

  dim(zarray)
  ## [1] 64 3007 3007

  format(object.size(zarray), units = "GB", standard = "SI")
  ## [1] "2.3 GB"
}

}
